feat: 메인페이지 스켈레톤 추가
