+++ src/components/AuctionList/AuctionCardList.tsx
@@ -26,7 +26,11 @@ function AuctionCardList({ sort, direction }: AuctionCardListProps) {
   });
 
   if (!data || isLoading) {
-    return <CardListSkeleton count={8} />;
+    return (
+      <div className={S.listWrapper}>
+        <CardListSkeleton count={8} />
+      </div>
+    );
   }
 
   return (
@@ -34,6 +38,7 @@ function AuctionCardList({ sort, direction }: AuctionCardListProps) {
       <ListLayout>
         {data.result_data.content.map(item => (
           <AuctionCard
+            categoryId={item.category_id}
             key={item.auction_id}
             id={item.auction_id}
             thumbnailImage={item.thumbnail_path}
